fastapi/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
celery[redis]
pytest==6.2.4
requests==2.25.1
Faker==8.14.0
python-dotenv==0.19.0